+++ luasrc/lldpd.lua
@@ -1,6 +1,11 @@
+--
+-- Copyright (c) 2018, Tano Systems. All Rights Reserved.
+-- Anton Kikin <a.kikin@tano-systems.com>
+--
+
 module "luci.lldpd"
 
-local app_version = "1.2.0-pre"
+local app_version = "1.2.0"
 
 function version()
 	return app_version