+++ CHANGELOG.md
@@ -9,10 +9,26 @@ All dates in this document are in `DD.MM.YYYY` format.
 ## [Unreleased]
 ### Added
 - This CHANGELOG.md file.
-- Added validators for parameters.
+- New screenshots.
+- Extra validators for parameters.
+- Local chassis information view.
+- The ability to configure the following new parameters:
+  * Port ID subtype applied to all interfaces (MAC or name of the interface)
+  * The destination MAC address used to send LLDPDU
+  * Receive-only mode (lldpd -r parameter)
+  * Interfaces to use for computing chassis ID (lldpd -C parameter)
+  * Filter neighbors (lldpd -H parameter)
+  * Disable LLDP-MED inventory TLV transmission (lldpd -i parameter)
+- Extra comments for existing parameters.
+
+### Changed
+- Improved tables appearance and usability. Allow to collapse/expand additional information
+  for each table row.
+- Merged status and config menu items as 3rd level pages under "Services"
+  main menu item.
 
 ### Fixed
-- Some HTML markup fixes.
+- HTML markup and JavaScript fixes.
 
 ## [Version 1.1.1] (18.09.2018)
 ### Fixed
@@ -27,7 +43,6 @@ All dates in this document are in `DD.MM.YYYY` format.
 
 Initial release
 
-[Unreleased]: https://github.com/tano-systems/luci-app-lldpd/tree/master
 [Version 1.1.1]: https://github.com/tano-systems/luci-app-lldpd/releases/tag/v1.1.1
 [Version 1.1.0]: https://github.com/tano-systems/luci-app-lldpd/releases/tag/v1.1.0
 [Version 1.0.0]: https://github.com/tano-systems/luci-app-lldpd/releases/tag/v1.0.0